requirements.txt
numpy
pandas
notebook
fastapi
openpyxl
mysql-connector-python
streamlit
uvicorn
python-dotenv